+++ deepcave/evaluators/epm/random_forest_surrogate.py
@@ -66,7 +66,7 @@ class RandomForestSurrogate(SurrogateModel):
         means, stds = self._model.predict(X)
         return means[:, 0], stds[:, 0]
 
-    def fit(self, X: np.ndarray, y: np.ndarray) -> None:
+    def _fit(self, X: np.ndarray, y: np.ndarray) -> None:
         """
         Train the surrogate model.
 
@@ -78,3 +78,16 @@ class RandomForestSurrogate(SurrogateModel):
             Corresponding target values.
         """
         self._model.train(X, y)
+
+    def fit(self, X: np.ndarray, y: np.ndarray) -> None:
+        """
+        Train the surrogate model.
+
+        Parameters
+        ----------
+        X : np.ndarray
+            Input data points.
+        y : np.ndarray
+            Corresponding target values.
+        """
+        self._fit(X, y)