+++ tests/test_evaluators/test_ablation.py
@@ -13,13 +13,108 @@
 # limitations under the License.
 
 
+from typing import Tuple
+
+import itertools
 import unittest
 
-import matplotlib.pyplot as plt
+import numpy as np
+from sympy import lambdify, symbols
 
 from deepcave.evaluators.ablation import Ablation as Evaluator
 from deepcave.runs import AbstractRun
 from deepcave.runs.converters.smac3v2 import SMAC3v2Run
+
+
+class PolynomialSurrogateModel:
+    """
+    This class is used to check if the ablation algorithm works
+    as intendet.
+
+    It generates a ground truth for testing purposes.
+    """
+
+    def __init__(self):
+        self.poly_func = None
+
+    def fit(self, X: np.ndarray, y: np.ndarray) -> None:
+        """
+        Since the model does not need fitting, this method does nothing.
+        It solely exists to fit the norm.
+        """
+        pass
+
+    def predict(self, X: np.ndarray) -> Tuple[np.ndarray, np.ndarray]:
+        """
+        Predict the output.
+        Since this is just a polynomial, the standard deviation is set to 0.
+
+        Parameters
+        ----------
+        X : np.ndarray
+            The coefficient mask to calcualate the polynomial.
+
+        Returns
+        -------
+        Tuple[np.ndarray, np.ndarray]
+            The result of the polynomial.
+        """
+        if self.poly_func is None:
+            self.polynomial(len(X[0]))
+            return
+
+        poly = self.poly_func(*X[0])
+
+        return np.array([poly]), np.array([0])
+
+    def polynomial(self, n: int) -> np.ndarray:
+        """
+        Get the multivariate basis polynomial with the given variables.
+
+        Parameters
+        ----------
+        n : int
+            The number of variables that should be used.
+
+        Returns
+        -------
+        np.ndarray
+            The fitting polynomial structure with random variables.
+        """
+        x = symbols(f"x1:{n+1}")
+        terms = []
+        a = []
+
+        count = 0
+
+        # This is fixed to 2 as to avoid too large computations
+        max_degree = 2
+        for d in range(1, max_degree + 1):
+            # All combinations of the variables are generated and added a coefficient.
+            for combo in itertools.combinations(x, r=d):
+                coeff = symbols(f"a{count}")
+                a.append(coeff)
+                term = coeff
+                for var in combo:
+                    term *= var
+                terms.append(term)
+                count += 1
+
+        # An additional a0 is added that stands alone
+        a0 = symbols("a0")
+        expr = a0 + sum(terms)
+
+        variables = [v for v in expr.free_symbols if str(v).startswith("x")]
+        variables = sorted(variables, key=lambda s: str(s))
+
+        np.random.seed(42)
+        input_values = np.random.uniform(1, 5, size=len(variables))
+        input_mapping = dict(zip(variables, input_values))
+        partial_expr = expr.subs(input_mapping)
+
+        coeff_symbols = sorted(partial_expr.free_symbols, key=lambda s: str(s))
+        self.poly_func = lambdify(coeff_symbols, partial_expr, modules="numpy")
+        return
 
 
 class TestAblation(unittest.TestCase):
@@ -61,25 +156,15 @@ class TestAblation(unittest.TestCase):
         budget = self.run.get_budget(0)
         objective = self.run.get_objective(0)
 
-        # Calculate
-        self.evaluator.calculate(objectives=objective, budget=budget, polynomial=True, degree=2)
-        self.poly_importance = self.evaluator.get_ablation_performances()
+        model = PolynomialSurrogateModel()
+        self.evaluator.calculate(objectives=objective, budget=budget, model=model)
 
-        self.evaluator.calculate(objectives=objective, budget=budget, polynomial=False)
-        self.rf_importance = self.evaluator.get_ablation_performances()
+        # Evaluate the final ablation path
+        performances = self.evaluator.get_ablation_performances()
 
-        keys = list(self.rf_importance.keys())
-        values1 = [v[0] for v in self.rf_importance.values()]
-        values2 = [v[0] for v in self.poly_importance.values()]
-
-        plt.plot(keys, values1, label="Random Forest Importances", marker="o")
-        plt.plot(keys, values2, label="Polynomial Importances", marker="s")
-
-        plt.xlabel("Hyperparameter")
-        plt.ylabel("Importance")
-        plt.legend()
-        plt.grid()
-        plt.show()
+        # Check if the performances are increasing
+        performances_list = [perf for perf, _ in performances.values()]
+        assert all(x <= y for x, y in zip(performances_list, performances_list[1:]))
 
 
 if __name__ == "__main__":